// frontend/src/app/generated-sources/openapi/model/showInformation.ts
/**
 * Ticketline API
 * Documentation for the Ticketline API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SeatWithBookingStatus } from './seatWithBookingStatus';
import { Sector } from './sector';
import { SeatingPlan } from './seatingPlan';


export interface ShowInformation { 
    seats: Array<SeatWithBookingStatus>;
    sectors: Array<Sector>;
    seatingPlan: SeatingPlan;
}